+++ src/export/web/transforms/types.ts
@@ -1,0 +1,15 @@
+import { GenericNode, MyST } from 'mystjs';
+
+export type Root = ReturnType<typeof MyST.prototype.parse>;
+
+export type Citations = {
+  order: string[];
+  data: Record<string, { html: string; number: number }>;
+};
+
+export type Footnotes = Record<string, GenericNode>;
+
+export type References = {
+  cite: Citations;
+  footnotes: Footnotes;
+};
